+++ traffic_emissions/core/operator_worker.py
@@ -6,9 +6,13 @@ from typing import List
 import shapely
 from climatoology.base.baseoperator import AoiProperties, BaseOperator, ComputationResources, _Artifact
 from climatoology.base.info import _Info
+from ohsome import OhsomeClient
 
 from traffic_emissions.components.traffic_emissions import (
+    get_district_summaries,
     get_emission_artifacts,
+    get_emission_chart_artifacts,
+    get_emission_sums,
     traffic_emissions,
 )
 from traffic_emissions.components.traffic_volume import build_traffic_volume_artifact, traffic_volume
@@ -19,6 +23,10 @@ log = logging.getLogger(__name__)
 
 
 class Operator(BaseOperator[ComputeInput]):
+    def __init__(self):
+        super().__init__()
+        self.ohsome = OhsomeClient()
+
     def info(self) -> _Info:
         return get_info()
 
@@ -29,13 +37,17 @@ class Operator(BaseOperator[ComputeInput]):
         aoi_properties: AoiProperties,
         params: ComputeInput,
     ) -> List[_Artifact]:
-        road_gdf = traffic_volume(aoi)
+        road_gdf = traffic_volume(aoi, self.ohsome)
         emissions_gdf = traffic_emissions(road_gdf)
+        emission_sums = get_emission_sums(emissions_gdf)
+        mean_df = get_district_summaries(emissions_gdf, aoi, self.ohsome)
 
         artifacts = []
         traffic_volume_artifact = build_traffic_volume_artifact(road_gdf, resources)
         artifacts.append(traffic_volume_artifact)
         emission_artifacts = get_emission_artifacts(emissions_gdf, resources)
         artifacts.extend(emission_artifacts)
+        emission_chart_artifacts = get_emission_chart_artifacts(mean_df, aoi_properties, emission_sums, resources)
+        artifacts.extend(emission_chart_artifacts)
 
         return artifacts